requirements.txt
typer>=0.12.0
httpx>=0.27.0
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
PyYAML>=6.0.1
google-api-python-client>=2.170.0
google-auth-oauthlib>=1.2.0
pytest>=8.0.0
